Ok, the app has reached the danger zone:
I want the app to be able to evolve these things and store cool ones

However, it's reached a point where it's very fun to just stare at
what's happening on the screen.

Fully random boards are created, and run. Once a loop or total death
happens, it restarts the process with another random seed.