All's well, but preparing to try a different DB approach
See, I found something that said you can store a vec in a db cell.

At first my gut said it's better to have cells be a join table. But now
I'm thinking I don't know about that. It miiiiiight be better form.. but
it's slower and more error prone, especially to inputs and reads when
there'll be many threads simultaneously using this thing. I'm thinking
many headaches will be solved if Vec<tuple> can really be saved in a
single cell.