Add two new evolution parameters!
starting_subdivisions and starting_subdiv_utilization

These regulate the size of the square the initial cells are placed in.

Also this commit brings about refactoring how boards are set up so
they're not so redundant, at the expense of a little uglier of a shape.
This is rust's composability without the ability to extend structs I
guess.